# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Fit&Grow
  description: This is the API for Fit&Grow, which is a sport activity monitoring system which allows the user to store his or her workouts, check his or her history and opt for achievements
  contact:
    name: Fit&Grow API team
    email: fitngrow2016@gmail.com
basePath: /api/v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /records:
    get:
      description: Returns all records from the system
      operationId: findRecords
      produces:
        - application/json
        - application/xml
        - text/xml
        - text/html
      responses:
        '200':
          description: An array of records
          schema:
            type: array
            items:
              $ref: '#/definitions/record'
        '500':
          description: Unexpected error
          schema:
              $ref: '#/definitions/error'
    post:
      description: Creates a new record in the system
      operationId: addRecord
      produces:
        - application/json
      parameters:
        - name: record
          in: body
          description: Record to add to the system
          required: true
          schema:
            $ref: '#/definitions/newRecord'
      responses:
        '201':
          description: Record response
          schema:
            $ref: '#/definitions/record'
        '409':
          description: Unexpected error
          schema:
            $ref: '#/definitions/error'
    delete:
      description: Deletes all records from the system
      operationId: deleteRecords
      responses:
        '200':
          description: Records deleted
        '500':
          description: Unexpected error
          schema:
              $ref: '#/definitions/error'
  /records/{_id}:
    get:
      description: Returns a record based on a single ID
      operationId: findRecordById
      parameters:
        - name: _id
          in: path
          description: ID of record to fetch
          required: true
          type: string
      responses:
        '200':
          description: Record response
          schema:
            $ref: '#/definitions/record'
        '404':
          description: There are not records in the system
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/error'
    put:
      description: Edits a record based on a single ID
      operationId: editRecordById
      parameters:
        - name: _id
          in: path
          description: ID of record to edit
          required: true
          type: string
        - name: record
          in: body
          description: Record to edit
          required: true
          schema:
            $ref: '#/definitions/record'
      responses:
        '200':
          description: Record updated
          schema:
            $ref: '#/definitions/record'
        '404':
          description: Record not found
          schema:
            $ref: '#/definitions/error'
        '409':
          description: There is not a record with that ID
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/error'
    delete:
      description: Deletes a single record based on the ID supplied
      operationId: deleteRecordById
      parameters:
        - name: _id
          in: path
          description: ID of record to delete
          required: true
          type: string
      responses:
        '200':
          description: Record deleted
        '404':
          description: Record not found
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/error'
  /records/user/{_id}:
    get:
      description: Returns the record of an user based on his or her ID
      operationId: findRecordByUserId
      parameters:
        - name: _id
          in: path
          description: ID of the user whose record is requested
          required: true
          type: string
      responses:
        '200':
          description: Record response
          schema:
            $ref: '#/definitions/prize'
        '404':
          description: The user does not have a record
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/error'
...
  /trainings:
    get:
      description: Returns all trainings from the system
      operationId: findTrainings
      produces:
        - application/json
        - application/xml
        - text/xml
        - text/html
      responses:
        '200':
          description: An array of trainings
          schema:
            type: array
            items:
              $ref: '#/definitions/training'
        '500':
          description: Unexpected error
          schema:
              $ref: '#/definitions/error'
    post:
      description: Creates a new training in the system
      operationId: addTraining
      produces:
        - application/json
      parameters:
        - name: training
          in: body
          description: Training to add to the system
          required: true
          schema:
            $ref: '#/definitions/newTraining'
      responses:
        '201':
          description: Training added
          schema:
            $ref: '#/definitions/training'
        '409':
          description: A training with the same ID already exists
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/error'
    delete:
      description: Deletes all trainings from the system
      operationId: deleteTrainings
      responses:
        '200':
          description: Trainings deleted
        '500':
          description: Error
          schema:
              $ref: '#/definitions/error'
  /trainings/{_id}:
    get:
      description: Returns a training based on a single ID
      operationId: findTrainingById
      parameters:
        - name: _id
          in: path
          description: ID of training to fetch
          required: true
          type: string
      responses:
        '200':
          description: Training response
          schema:
            $ref: '#/definitions/training'
        '404':
          description: Training not found
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/error'
    put:
      description: Edits a training based on a single ID
      operationId: editTrainingById
      parameters:
        - name: _id
          in: path
          description: ID of training to edit
          required: true
          type: string
        - name: training
          in: body
          description: Training to edit
          required: true
          schema:
            $ref: '#/definitions/training'
      responses:
        '200':
          description: Training updated
          schema:
            $ref: '#/definitions/training'
        '404':
          description: Training not found
          schema:
            $ref: '#/definitions/error'
        '409':
          description: The identifier is not correct
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/error'
    delete:
      description: Deletes a single training based on the ID supplied
      operationId: deleteTrainingById
      parameters:
        - name: _id
          in: path
          description: ID of training to delete
          required: true
          type: string
      responses:
        '200':
          description: Training deleted
        '404':
          description: Training not found
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/error'
  /trainings/user/{_id}:
    get:
      description: Returns the traininf of an user based on his or her ID
      operationId: findTrainingsdByUserId
      parameters:
        - name: _id
          in: path
          description: ID of the user whose trainings are requested
          required: true
          type: string
      responses:
        '200':
          description: Training response
          schema:
            $ref: '#/definitions/prize'
        '404':
          description: Training not found
          schema:
            $ref: '#/definitions/error'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/error'
...
definitions:
...
  prize:
    type: object
    required:
      - _id
      - idUser
      - name
      - description
      - code
      - type
    properties:
      _id:
        type: string
      idUser:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
      code:
        type: string
      type:
        type: string
...
  record:
    type: object
    required:
      - _id
      - idUser
      - distance
      - sessions
      - averageDistance
      - calories
      - totalTime
    properties:
      _id:
        type: string
      idUser:
        type: integer
        format: int64
      distance:
        type: number
        format: double
      sessions:
        type: integer
        format: int64
      averageDistance:
        type: number
        format: double
      calories:
        type: number
        format: double
      totalTime:
        type: number
        format: double
  newRecord:
    type: object
    required:
      - idUser
      - distance
      - sessions
      - averageDistance
      - calories
      - totalTime
    properties:
      _id:
        type: string
      idUser:
        type: integer
        format: int64
      distance:
        type: number
        format: double
      sessions:
        type: integer
        format: int64
      averageDistance:
        type: number
        format: double
      calories:
        type: number
        format: double
      totalTime:
        type: number
        format: double
...
  training:
    type: object
    required:
      - _id
      - idUser
      - idSport
      - start
      - end
      - calories
      - distance
    properties:
      _id:
        type: string
      idUser:
        type: integer
        format: int64
      idSport:
        type: integer
        format: int64
      start:
        type: string
      end:
        type: string
      calories:
        type: number
        format: double
      distance:
        type: number
        format: double
  newTraining:
    type: object
    required:
      - idUser
      - idSport
      - start
      - end
      - calories
      - distance
    properties:
      _id:
        type: string
      idUser:
        type: integer
        format: int64
      idSport:
        type: integer
        format: int64
      start:
        type: string
      end:
        type: string
      calories:
        type: number
        format: double
      distance:
        type: number
        format: double
...
  error:
    type: integer
    format: int64